Pull request #3821: [BLUETOOTH-17040] Fix MISRA issues by adding cast and removing redundant code
Merge in CONNINT/bluetooth_public from bugfix/BLUETOOTH-17040-fix-misra-issues-in-ccc-key-device-cs to develop/main

Squashed commit of the following:

commit 50b899cd91bd72d47a01e3375a41fbe37c1e1acf
Author: Andrei Istodorescu <andrei.istodorescu@nxp.com>
Date:   Wed Feb 18 21:33:17 2026 +0200

    [BLUETOOTH-17040] Fix MISRA issues by adding cast and removing redundant code

+++ application/common/lcl/btcs_server.c
@@ -886,7 +886,7 @@ static void parseSubeventData
 
             case (uint8_t)gCsStepMode3_c:
             {
-                if ((dataCopiedLen + (mpRangingData[deviceId].pCfg->numAntennaPaths*gTone_PCTSize_c) +
+                if ((dataCopiedLen + ((uint16_t)mpRangingData[deviceId].pCfg->numAntennaPaths*gTone_PCTSize_c) +
                      2U + gMode1DataSize_c)> maxDataLen)
                 {
                     /* There is no more room for this step - break loop */
